Refactor VariantTasksList
